Enhance email update response and add validation for unique phone numbers in user update views

+++ accounts/views/user_update_views.py
@@ -4,7 +4,7 @@ from rest_framework.views import APIView
 from rest_framework import permissions
 from django.core.cache import cache
 
-from accounts.models import UserProfile
+from accounts.models import UserProfile, User
 from accounts.otp import *
 from accounts.serializers.user_update_serializers import *
 from accounts.serializers.auth_serializers import RequestOTPSerializer, VerifyOTPRequestSerializer
@@ -37,9 +37,12 @@ class SetEmailView(APIView):
         user=request.user
         serializer = self.serializer_class(user, data=request.data, partial=True)
 
-        if user.email is None and serializer.is_valid():
+        if serializer.is_valid():
             serializer.save()
-            return Response(serializer.data, status=status.HTTP_200_OK)
+            return Response(
+                {"message": "Email updated successfully.", "data": serializer.data},
+                status=status.HTTP_200_OK
+            )
         return Response(serializer.errors, status=status.HTTP_400_BAD_REQUEST)
 
 
@@ -54,15 +57,17 @@ class ChangeNumberRequestView(APIView):
         if serializer.is_valid():
             data = serializer.validated_data
             new_number = data['number']
-            if user.number != new_number:
+            if not User.objects.filter(number=new_number).exists():
+                if user.number != new_number:
 
-                delete_otp_change_number(user.id)
-                cache.delete(f"new_number_{user.id}")
-                otp = generate_otp_change_number(user.id)
-                print(f'Your OTP is: {otp}')
-                cache.set(f"new_number_{user.id}", new_number, timeout=OTP_TIMEOUT)
-                return Response({"detail": "OTP sent successfully."}, status=status.HTTP_200_OK)
-            return Response({"detail": "New number is same as old number."}, status=status.HTTP_400_BAD_REQUEST)
+                    delete_otp_change_number(user.id)
+                    cache.delete(f"new_number_{user.id}")
+                    otp = generate_otp_change_number(user.id)
+                    print(f'Your OTP is: {otp}')
+                    cache.set(f"new_number_{user.id}", new_number, timeout=OTP_TIMEOUT)
+                    return Response({"detail": "OTP sent successfully."}, status=status.HTTP_200_OK)
+                return Response({"detail": "New number is same as old number."}, status=status.HTTP_400_BAD_REQUEST)
+            return Response({"detail": "Number already exists."}, status=status.HTTP_400_BAD_REQUEST)
         
         return Response(serializer.errors, status=status.HTTP_400_BAD_REQUEST)
